requirements.txt
boto3==1.24.82
botocore==1.27.82
Cerberus==1.3.4
paramiko==3.0.0
psycopg2==2.9.5
python-dotenv==1.0.0
ruamel.base==1.0.0
sshtunnel==0.4.0
